Add documentField and create simple saved card component
