+++ manifest.json
@@ -18,7 +18,8 @@
   "dependencies": {
     "vtex.styleguide": "9.x",
     "vtex.order-manager": "0.x",
-    "vtex.order-payment": "0.x"
+    "vtex.order-payment": "0.x",
+    "vtex.document-field": "0.x"
   },
   "$schema": "https://raw.githubusercontent.com/vtex/node-vtex-api/master/gen/manifest.schema"
 }
