+++ messages/en.json
@@ -1,4 +1,6 @@
 {
   "checkout-payment.button.save": "Save new card",
-  "checkout-payment.input.document": "Cardholder document"
+  "checkout-payment.input.document": "Cardholder document",
+  "checkout-payment.input.requiredField": "This field is required.",
+  "checkout-payment.input.invalidDigits": "Invalid Digits"
 }
