+++ messages/es.json
@@ -1,4 +1,6 @@
 {
   "checkout-payment.button.save": "Guardar tarjeta",
-  "checkout-payment.input.document": "Documento del titular de la tarjeta"
+  "checkout-payment.input.document": "Documento del titular de la tarjeta",
+  "checkout-payment.input.requiredField": "Este campo es obligatorio.",
+  "checkout-payment.input.invalidDigits": "Dígitos inválidos"
 }
