+++ messages/pt.json
@@ -1,4 +1,6 @@
 {
   "checkout-payment.button.save": "Salvar novo cartão",
-  "checkout-payment.input.document": "Documento do titular do cartão"
+  "checkout-payment.input.document": "Documento do titular do cartão",
+  "checkout-payment.input.requiredField": "Campo obrigatório.",
+  "checkout-payment.input.invalidDigits": "Dígitos inválidos"
 }
