+++ react/typings/vtex.document-field.d.ts
@@ -1,0 +1,1 @@
+declare module 'vtex.document-field'
